backend/requirements.txt
annotated-types==0.6.0
asgiref==3.8.1
boto3==1.35.44
botocore==1.35.44
cffi==1.17.1
contextlib2==21.6.0
cryptography==43.0.1
Django==5.0.4
django-cors-headers==4.4.0
django-debug-toolbar==4.3.0
django-environ==0.11.2
django-ninja==1.3.0
django-ninja-extra==0.21.4
django-ninja-jwt==5.3.4
dnspython==2.6.1
email_validator==2.2.0
gunicorn==22.0.0
idna==3.8
injector==0.22.0
jmespath==1.0.1
markdown-it-py==3.0.0
mdurl==0.1.2
numpy==2.1.1
packaging==24.1
pillow==10.3.0
psycopg==3.2.1
psycopg-binary==3.2.1
psycopg-pool==3.2.3
pycparser==2.22
pydantic==2.6.4
pydantic_core==2.16.3
Pygments==2.17.2
PyJWT==2.9.0
python-dateutil==2.9.0.post0
rich==13.7.1
s3transfer==0.10.3
setuptools==68.2.2
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.3
wheel==0.41.2
